OGS6_pvd

Constructor for the OGS6_pvd base class

Active bindings

pvd_path

Getter for private parameter '.pvd_path'

datasets

Getter for private parameter '.datasets'

vtu_paths

Getter for `datasets` `file`

abs_vtu_paths

Gets absolute .vtu paths, e.g. `dirname(pvd_path)` + `datasets` `file`

last_timestep

Gets last timestep

timesteps

Gets timesteps from private parameter `datasets`

OGS6_vtus

Getter for private parameter `.OGS6_vtus`

Methods

Public methods

OGS6_pvd$new()

OGS6_pvd$print()

OGS6_pvd$vtu_by_timestep()

OGS6_pvd$timestep_by_vtu()

OGS6_pvd$get_point_data_at()

OGS6_pvd$get_point_data()

OGS6_pvd$get_cell_data()

OGS6_pvd$clone()

Method new()

Creates new OGS6_pvd object

Usage

OGS6_pvd$new(pvd_path)

Arguments

pvd_path

string: Path to .pvd file

Method print()

Overrides default printing behaviour

Usage

OGS6_pvd$print()

Method vtu_by_timestep()

Returns .vtu path for specified timestep

Usage

OGS6_pvd$vtu_by_timestep(timestep)

Arguments

timestep

string: Timestep

Method timestep_by_vtu()

Returns timestep for specified .vtu path

Usage

OGS6_pvd$timestep_by_vtu(vtu_path)

Arguments

vtu_path

string: .vtu path

Method get_point_data_at()

Returns a tibble containing point data

Usage

OGS6_pvd$get_point_data_at(
  coordinates,
  keys,
  start_at_timestep,
  end_at_timestep
)

Arguments

coordinates

list(numeric): List of coordinates (a coordinate is a numeric vector of length 3)

keys

character: Optional: `Name` attributes of `DataArray` elements. Defaults to all.

start_at_timestep

number: Optional: Timestep to start at. Defaults to first timestep.

end_at_timestep

number: Optional: Timestep to end at. Defaults to last timestep.

Method get_point_data()

Returns a tibble containing point data

Usage

OGS6_pvd$get_point_data(point_ids, keys, start_at_timestep, end_at_timestep)

Arguments

point_ids

numeric: Optional: Point IDs. Defaults to all.

keys

character: Optional: `Name` attributes of `DataArray` elements. Defaults to all.

start_at_timestep

number: Optional: Timestep to start at. Defaults to first timestep.

end_at_timestep

number: Optional: Timestep to end at. Defaults to last timestep.

Method get_cell_data()

Returns a tibble containing cell data

Usage

OGS6_pvd$get_cell_data(cell_ids, keys, start_at_timestep, end_at_timestep)

Arguments

cell_ids

numeric: Optional: Cell IDs. Defaults to all.

keys

character: Optional: `Name` attributes of `DataArray` elements. Defaults to all.

start_at_timestep

number: Optional: Timestep to start at. Defaults to first timestep.

end_at_timestep

number: Optional: Timestep to end at. Defaults to last timestep.

Method clone()

The objects of this class are cloneable with this method.

Usage

OGS6_pvd$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.